feat: add better support for the celestial_frame-1.2.0 schema
internally it's called `frame_celestial` instead of `celestial_frame` in
order to consistently name all frame types like
`asdf_gwcs_frame_*_t`

refactor: more refactoring of frame parsing to make it reusable across
different GWCS frame types, though I'm still not sure I'm happy with how
the structs are laid out

gh-66

+++ include/asdf/gwcs/frame2d.h
@@ -16,7 +16,7 @@ ASDF_BEGIN_DECLS
 typedef struct {
     asdf_gwcs_frame_t base;
     const char *axes_names[2];
-    uint8_t axes_order[2];
+    uint32_t axes_order[2];
     // TODO: Should be an asdf_unit_t but right now that is just a string
     const char *unit[2];
 } asdf_gwcs_frame2d_t;
